requirements.txt
aiohttp
aiosqlite
discord.py
pytest